Fix image links and use icon-24 instead of icon-64
The link for the github images are broken due to which the images are not rendered in the mkdocs page. I have fixed that with a sed command. Also I felt that the the icon-64 image so I changed to icon-24. I have used a sed command to do that too.